+++ dwarfexplorer/QHexView/document/qhexcursor.h
@@ -7,7 +7,7 @@
 #define HEX_LINE_LAST_COLUMN (HEX_LINE_LENGTH - 1)
 
 struct QHexPosition {
-    int line, column, nibbleindex;
+    uint64_t line, column, nibbleindex;
 
     QHexPosition() = default;
     QHexPosition(const QHexPosition&) = default;
@@ -19,10 +19,10 @@ struct QHexPosition {
         return *this;
     }
 
-    int offset() const { return (line * HEX_LINE_LENGTH) + column; }
-    int operator-(const QHexPosition& rhs) const { return this->offset() - rhs.offset(); }
-    bool operator==(const QHexPosition& rhs) const { return (line == rhs.line) && (column == rhs.column) && (nibbleindex == rhs.nibbleindex); }
-    bool operator!=(const QHexPosition& rhs) const { return (line != rhs.line) || (column != rhs.column) || (nibbleindex != rhs.nibbleindex); }
+    uint64_t offset() const { return (line * HEX_LINE_LENGTH) + column; }
+    uint64_t operator-(const QHexPosition& rhs) const { return this->offset() - rhs.offset(); }
+    bool     operator==(const QHexPosition& rhs) const { return (line == rhs.line) && (column == rhs.column) && (nibbleindex == rhs.nibbleindex); }
+    bool     operator!=(const QHexPosition& rhs) const { return (line != rhs.line) || (column != rhs.column) || (nibbleindex != rhs.nibbleindex); }
 };
 
 class QHexCursor : public QObject
@@ -40,26 +40,26 @@ class QHexCursor : public QObject
         const QHexPosition& selectionEnd() const;
         const QHexPosition& position() const;
         InsertionMode insertionMode() const;
-        int selectionLength() const;
-        int currentLine() const;
-        int currentColumn() const;
-        int currentNibble() const;
-        int selectionLine() const;
-        int selectionColumn() const;
-        int selectionNibble() const;
+        uint64_t selectionLength() const;
+        uint64_t currentLine() const;
+        uint64_t currentColumn() const;
+        uint64_t currentNibble() const;
+        uint64_t selectionLine() const;
+        uint64_t selectionColumn() const;
+        uint64_t selectionNibble() const;
         bool atEnd() const;
-        bool isLineSelected(int line) const;
+        bool isLineSelected(uint64_t line) const;
         bool hasSelection() const;
         void clearSelection();
 
     public:
         void moveTo(const QHexPosition& pos);
-        void moveTo(int line, int column, int nibbleindex = 1);
-        void moveTo(int offset);
+        void moveTo(uint64_t line, uint64_t column, uint64_t nibbleindex = 1);
+        void moveTo(uint64_t offset);
         void select(const QHexPosition& pos);
-        void select(int line, int column, int nibbleindex = 1);
-        void select(int length);
-        void selectOffset(int offset, int length);
+        void select(uint64_t line, uint64_t column, uint64_t nibbleindex = 1);
+        void select(uint64_t length);
+        void selectOffset(uint64_t offset, uint64_t length);
         void setInsertionMode(InsertionMode mode);
         void switchInsertionMode();
